Deprecate ADD_TRAILING_SLASH_TO_LOCATION instead of removing it
Keep the feature flag registered as @Deprecated(forRemoval) so its catalog
config key (polaris.config.add-trailing-slash-to-location) stays on the
ReservedProperties allowlist; removing the declaration made catalog
create/update reject the leftover key with a 400. The flag remains unread at
runtime and the production-readiness warning still fires when it is set to
false.

Normalize the namespace base location in setProperties via
StorageLocation.ensureTrailingSlash so property updates honor the same
trailing-slash invariant as createNamespaceInternal.

Move the QueryGenerator trailing-slash regression test into the JDBC-only
overlap test, since only the relational-jdbc optimized sibling check exercises
that path; the in-memory and NoSQL profiles gave false confidence.
