+++ runtime/service/src/test/java/org/apache/polaris/service/catalog/iceberg/AbstractLocalIcebergCatalogOverlapTest.java
@@ -19,7 +19,6 @@
 package org.apache.polaris.service.catalog.iceberg;
 
 import static org.apache.polaris.service.admin.PolarisAuthzTestBase.SCHEMA;
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.when;
 
@@ -49,15 +48,10 @@ import org.apache.polaris.core.config.RealmConfig;
 import org.apache.polaris.core.context.CallContext;
 import org.apache.polaris.core.entity.CatalogEntity;
 import org.apache.polaris.core.entity.PolarisEntity;
-import org.apache.polaris.core.entity.PolarisEntityCore;
-import org.apache.polaris.core.entity.PolarisEntitySubType;
-import org.apache.polaris.core.entity.PolarisEntityType;
-import org.apache.polaris.core.entity.table.IcebergTableLikeEntity;
 import org.apache.polaris.core.identity.provider.ServiceIdentityProvider;
 import org.apache.polaris.core.persistence.MetaStoreManagerFactory;
 import org.apache.polaris.core.persistence.PolarisMetaStoreManager;
 import org.apache.polaris.core.persistence.bootstrap.RootCredentialsSet;
-import org.apache.polaris.core.persistence.dao.entity.EntityResult;
 import org.apache.polaris.core.persistence.resolver.ResolutionManifestFactory;
 import org.apache.polaris.core.persistence.resolver.ResolverFactory;
 import org.apache.polaris.core.storage.PolarisStorageIntegrationProvider;
@@ -126,9 +120,9 @@ public abstract class AbstractLocalIcebergCatalogOverlapTest {
 
   private LocalIcebergCatalog catalog;
   private String realmName;
-  private PolarisCallContext polarisContext;
+  protected PolarisCallContext polarisContext;
   private InMemoryFileIO fileIO;
-  private PolarisEntity catalogEntity;
+  protected PolarisEntity catalogEntity;
 
   @BeforeAll
   public static void setUpMocks() {
@@ -300,98 +294,5 @@ public abstract class AbstractLocalIcebergCatalogOverlapTest {
         .isInstanceOf(ForbiddenException.class)
         .hasMessageContaining("Unable to create entity at location")
         .hasMessageContaining("conflicts with existing table or namespace");
-  }
-
-  @Test
-  public void testParentPrefixOverlapWithTrailingSlashMismatch() {
-    // The catalog now always stores locations with a trailing slash
-    // (ADD_TRAILING_SLASH_TO_LOCATION was removed). Simulate legacy slash-less data
-    // by rewriting the stored parent location directly, then verify the optimized sibling check
-    // still detects overlaps against it: ancestor equality terms used to be slash-terminated only,
-    // so a slash-less location_without_scheme was missed.
-    Namespace ns = Namespace.of("ns-for-trailing-slash-overlap");
-    catalog().createNamespace(ns);
-
-    TableIdentifier parentTable = TableIdentifier.of(ns, "trailing-parent");
-    String parentLoc = STORAGE_LOCATION + "/trailing-overlap/parent";
-    assertThat(parentLoc).doesNotEndWith("/");
-    catalog().buildTable(parentTable, SCHEMA).withLocation(parentLoc).create();
-
-    stripStoredBaseLocationTrailingSlash(ns, parentTable);
-
-    // Guardrail: the stored location must be slash-less so overlap checks exercise the
-    // non-slash-terminated location_without_scheme path that QueryGenerator must handle.
-    assertThat(storedTableBaseLocation(ns, parentTable))
-        .as("parent must remain slash-less so overlap uses non-slash-terminated stored location")
-        .isEqualTo(parentLoc)
-        .doesNotEndWith("/");
-
-    // Creating a child table under the slash-less parent location must be rejected.
-    TableIdentifier childTable = TableIdentifier.of(ns, "trailing-child");
-    String childLoc = STORAGE_LOCATION + "/trailing-overlap/parent/child";
-    assertThatThrownBy(
-            () -> catalog().buildTable(childTable, SCHEMA).withLocation(childLoc).create())
-        .isInstanceOf(ForbiddenException.class)
-        .hasMessageContaining("Unable to create entity at location")
-        .hasMessageContaining("conflicts with existing table or namespace");
-
-    // The same must hold when the slash-terminated form of the same prefix is used.
-    TableIdentifier siblingTable = TableIdentifier.of(ns, "trailing-sibling");
-    String siblingLoc = STORAGE_LOCATION + "/trailing-overlap/parent/";
-    assertThatThrownBy(
-            () -> catalog().buildTable(siblingTable, SCHEMA).withLocation(siblingLoc).create())
-        .isInstanceOf(ForbiddenException.class)
-        .hasMessageContaining("Unable to create entity at location")
-        .hasMessageContaining("conflicts with existing table or namespace");
-  }
-
-  private String storedTableBaseLocation(Namespace ns, TableIdentifier table) {
-    return IcebergTableLikeEntity.of(readStoredTableLike(ns, table)).getBaseLocation();
-  }
-
-  /**
-   * Rewrites the stored base location of {@code table} without its trailing slash, simulating data
-   * written before Polaris always appended one.
-   */
-  private void stripStoredBaseLocationTrailingSlash(Namespace ns, TableIdentifier table) {
-    PolarisEntity tableEntity = readStoredTableLike(ns, table);
-    IcebergTableLikeEntity tableLike = IcebergTableLikeEntity.of(tableEntity);
-    String baseLocation = tableLike.getBaseLocation();
-    assertThat(baseLocation).endsWith("/");
-    IcebergTableLikeEntity stripped =
-        new IcebergTableLikeEntity.Builder(tableLike)
-            .setBaseLocation(baseLocation.substring(0, baseLocation.length() - 1))
-            .build();
-    EntityResult result =
-        metaStoreManager.updateEntityPropertiesIfNotChanged(
-            polarisContext, tableLikeCatalogPath(ns), stripped);
-    assertThat(result.isSuccess()).isTrue();
-  }
-
-  private PolarisEntity readStoredTableLike(Namespace ns, TableIdentifier table) {
-    EntityResult result =
-        metaStoreManager.readEntityByName(
-            polarisContext,
-            tableLikeCatalogPath(ns),
-            PolarisEntityType.TABLE_LIKE,
-            PolarisEntitySubType.ICEBERG_TABLE,
-            table.name());
-    assertThat(result.isSuccess()).isTrue();
-    return PolarisEntity.of(result.getEntity());
-  }
-
-  private List<PolarisEntityCore> tableLikeCatalogPath(Namespace ns) {
-    assertThat(ns.length()).as("test namespaces must be single-level").isEqualTo(1);
-    EntityResult result =
-        metaStoreManager.readEntityByName(
-            polarisContext,
-            List.of(PolarisEntity.toCore(catalogEntity)),
-            PolarisEntityType.NAMESPACE,
-            PolarisEntitySubType.NULL_SUBTYPE,
-            ns.level(0));
-    assertThat(result.isSuccess()).isTrue();
-    return List.of(
-        PolarisEntity.toCore(catalogEntity),
-        PolarisEntity.toCore(PolarisEntity.of(result.getEntity())));
   }
 }
